[popups] Reuse Viewport logic across components
