+++ packages/react/src/preview-card/viewport/PreviewCardViewport.tsx
@@ -1,20 +1,12 @@
 'use client';
 import * as React from 'react';
-import { inertValue } from '@base-ui/utils/inertValue';
-import { useAnimationFrame } from '@base-ui/utils/useAnimationFrame';
-import { usePreviousValue } from '@base-ui/utils/usePreviousValue';
-import { useIsoLayoutEffect } from '@base-ui/utils/useIsoLayoutEffect';
-import { useStableCallback } from '@base-ui/utils/useStableCallback';
 import { usePreviewCardRootContext } from '../root/PreviewCardContext';
 import { usePreviewCardPositionerContext } from '../positioner/PreviewCardPositionerContext';
 import { BaseUIComponentProps } from '../../utils/types';
-import { useAnimationsFinished } from '../../utils/useAnimationsFinished';
-import { usePopupAutoResize } from '../../utils/usePopupAutoResize';
 import { useRenderElement } from '../../utils/useRenderElement';
 import { StateAttributesMapping } from '../../utils/getStateAttributesProps';
-import { Dimensions } from '../../floating-ui-react/types';
 import { PreviewCardViewportCssVars } from './PreviewCardViewportCssVars';
-import { useDirection } from '../../direction-provider';
+import { usePopupViewport } from '../../utils/usePopupViewport';
 
 const stateAttributesMapping: StateAttributesMapping<PreviewCardViewport.State> = {
   activationDirection: (value) =>
@@ -40,191 +32,16 @@ export const PreviewCardViewport = React.forwardRef(function PreviewCardViewport
   const { render, className, children, ...elementProps } = componentProps;
   const store = usePreviewCardRootContext();
   const positioner = usePreviewCardPositionerContext();
-  const direction = useDirection();
 
-  const activeTrigger = store.useState('activeTriggerElement');
-  const activeTriggerId = store.useState('activeTriggerId');
-  const payload = store.useState('payload');
-  const open = store.useState('open');
   const instantType = store.useState('instantType');
-  const mounted = store.useState('mounted');
-  const popupElement = store.useState('popupElement');
-  const positionerElement = store.useState('positionerElement');
 
-  const previousActiveTrigger = usePreviousValue(open ? activeTrigger : null);
-  // Remount current content on trigger changes (and once more when payload lags) to avoid DOM reuse flashes.
-  // The key bumps immediately on trigger switches, then again if the payload arrives on a later render.
-  const currentContentKey = usePopupContentKey(activeTriggerId, payload);
-
-  const capturedNodeRef = React.useRef<HTMLElement | null>(null);
-  const [previousContentNode, setPreviousContentNode] = React.useState<HTMLElement | null>(null);
-
-  const [newTriggerOffset, setNewTriggerOffset] = React.useState<Offset | null>(null);
-
-  const currentContainerRef = React.useRef<HTMLDivElement>(null);
-  const previousContainerRef = React.useRef<HTMLDivElement>(null);
-
-  const onAnimationsFinished = useAnimationsFinished(currentContainerRef, true, false);
-  const cleanupFrame = useAnimationFrame();
-
-  const [previousContentDimensions, setPreviousContentDimensions] = React.useState<{
-    width: number;
-    height: number;
-  } | null>(null);
-
-  const [showStartingStyleAttribute, setShowStartingStyleAttribute] = React.useState(false);
-
-  useIsoLayoutEffect(() => {
-    store.set('hasViewport', true);
-    return () => {
-      store.set('hasViewport', false);
-    };
-  }, [store]);
-
-  const handleMeasureLayout = useStableCallback(() => {
-    currentContainerRef.current?.style.setProperty('animation', 'none');
-    currentContainerRef.current?.style.setProperty('transition', 'none');
-
-    previousContainerRef.current?.style.setProperty('display', 'none');
-  });
-
-  const handleMeasureLayoutComplete = useStableCallback((previousDimensions: Dimensions | null) => {
-    currentContainerRef.current?.style.removeProperty('animation');
-    currentContainerRef.current?.style.removeProperty('transition');
-
-    previousContainerRef.current?.style.removeProperty('display');
-
-    if (previousDimensions) {
-      setPreviousContentDimensions(previousDimensions);
-    }
-  });
-
-  const lastHandledTriggerRef = React.useRef<Element | null>(null);
-
-  useIsoLayoutEffect(() => {
-    // When a trigger changes, set the captured children HTML to state,
-    // so we can render both new and old content.
-    if (
-      activeTrigger &&
-      previousActiveTrigger &&
-      activeTrigger !== previousActiveTrigger &&
-      lastHandledTriggerRef.current !== activeTrigger &&
-      capturedNodeRef.current
-    ) {
-      setPreviousContentNode(capturedNodeRef.current);
-      setShowStartingStyleAttribute(true);
-
-      // Calculate the relative position between the previous and new trigger,
-      // so we can pass it to the style hook for animation purposes.
-      const offset = calculateRelativePosition(previousActiveTrigger, activeTrigger);
-      setNewTriggerOffset(offset);
-
-      cleanupFrame.request(() => {
-        cleanupFrame.request(() => {
-          setShowStartingStyleAttribute(false);
-          onAnimationsFinished(() => {
-            setPreviousContentNode(null);
-            setPreviousContentDimensions(null);
-            capturedNodeRef.current = null;
-          });
-        });
-      });
-
-      lastHandledTriggerRef.current = activeTrigger;
-    }
-  }, [
-    activeTrigger,
-    previousActiveTrigger,
-    previousContentNode,
-    onAnimationsFinished,
-    cleanupFrame,
-  ]);
-
-  // Capture a clone of the current content DOM subtree when not transitioning.
-  // We can't store previous React nodes as they may be stateful; instead we capture DOM clones for visual continuity.
-  useIsoLayoutEffect(() => {
-    // When a transition is in progress, we store the next content in capturedNodeRef.
-    // This handles the case where the trigger changes multiple times before the transition finishes.
-    // We want to always capture the latest content for the previous snapshot.
-    // So clicking quickly on T1, T2, T3 will result in the following sequence:
-    // 1. T1 -> T2: previousContent = T1, currentContent = T2
-    // 2. T2 -> T3: previousContent = T2, currentContent = T3
-    const source = currentContainerRef.current;
-    if (!source) {
-      return;
-    }
-
-    const wrapper = document.createElement('div');
-    for (const child of Array.from(source.childNodes)) {
-      wrapper.appendChild(child.cloneNode(true));
-    }
-
-    capturedNodeRef.current = wrapper;
-  });
-
-  const isTransitioning = previousContentNode != null;
-  let childrenToRender: React.ReactNode;
-  if (!isTransitioning) {
-    childrenToRender = (
-      <div data-current ref={currentContainerRef} key={currentContentKey}>
-        {children}
-      </div>
-    );
-  } else {
-    childrenToRender = (
-      <React.Fragment>
-        <div
-          data-previous
-          inert={inertValue(true)}
-          ref={previousContainerRef}
-          style={
-            {
-              [PreviewCardViewportCssVars.popupWidth]: `${previousContentDimensions?.width}px`,
-              [PreviewCardViewportCssVars.popupHeight]: `${previousContentDimensions?.height}px`,
-              position: 'absolute',
-            } as React.CSSProperties
-          }
-          key={'previous'}
-          data-ending-style={showStartingStyleAttribute ? undefined : ''}
-        />
-        <div
-          data-current
-          ref={currentContainerRef}
-          key={currentContentKey}
-          data-starting-style={showStartingStyleAttribute ? '' : undefined}
-        >
-          {children}
-        </div>
-      </React.Fragment>
-    );
-  }
-
-  // When previousContentNode is present, imperatively populate the previous container with the cloned children.
-  useIsoLayoutEffect(() => {
-    const container = previousContainerRef.current;
-    if (!container || !previousContentNode) {
-      return;
-    }
-
-    container.replaceChildren(...Array.from(previousContentNode.childNodes));
-  }, [previousContentNode]);
-
-  usePopupAutoResize({
-    popupElement,
-    positionerElement,
-    mounted,
-    content: payload,
-    onMeasureLayout: handleMeasureLayout,
-    onMeasureLayoutComplete: handleMeasureLayoutComplete,
+  const { children: childrenToRender, state } = usePopupViewport({
+    store,
     side: positioner.side,
-    direction,
+    instantType,
+    cssVars: PreviewCardViewportCssVars,
+    children,
   });
-
-  const state: PreviewCardViewport.State = {
-    activationDirection: getActivationDirection(newTriggerOffset),
-    transitioning: isTransitioning,
-    instant: instantType,
-  };
 
   return useRenderElement('div', componentProps, {
     state,
@@ -250,106 +67,4 @@ export namespace PreviewCardViewport {
     transitioning: boolean;
     instant: 'dismiss' | 'focus' | undefined;
   }
-}
-
-type Offset = {
-  horizontal: number;
-  vertical: number;
-};
-
-/**
- * Returns a string describing the provided offset.
- * It describes both the horizontal and vertical offset, separated by a space.
- *
- * @param offset
- */
-function getActivationDirection(offset: Offset | null): string | undefined {
-  if (!offset) {
-    return undefined;
-  }
-
-  return `${getValueWithTolerance(offset.horizontal, 5, 'right', 'left')} ${getValueWithTolerance(offset.vertical, 5, 'down', 'up')}`;
-}
-
-/**
- * Returns a label describing the value (positive/negative) treating values
- * within tolerance as zero.
- *
- * @param value Value to check
- * @param tolerance Tolerance to treat the value as zero.
- * @param positiveLabel
- * @param negativeLabel
- * @returns If 0 < abs(value) < tolerance, returns an empty string. Otherwise returns positiveLabel or negativeLabel.
- */
-function getValueWithTolerance(
-  value: number,
-  tolerance: number,
-  positiveLabel: string,
-  negativeLabel: string,
-) {
-  if (value > tolerance) {
-    return positiveLabel;
-  }
-
-  if (value < -tolerance) {
-    return negativeLabel;
-  }
-
-  return '';
-}
-
-/**
- * Calculates the relative position between centers of two elements.
- */
-function calculateRelativePosition(from: Element, to: Element): Offset {
-  const fromRect = from.getBoundingClientRect();
-  const toRect = to.getBoundingClientRect();
-
-  const fromCenter = {
-    x: fromRect.left + fromRect.width / 2,
-    y: fromRect.top + fromRect.height / 2,
-  };
-  const toCenter = {
-    x: toRect.left + toRect.width / 2,
-    y: toRect.top + toRect.height / 2,
-  };
-
-  return {
-    horizontal: toCenter.x - fromCenter.x,
-    vertical: toCenter.y - fromCenter.y,
-  };
-}
-
-/**
- * Returns a key that forces remounting content when triggers change or a payload is updated.
- */
-function usePopupContentKey(activeTriggerId: string | null, payload: unknown): string {
-  const [contentKey, setContentKey] = React.useState(0);
-  const previousActiveTriggerIdRef = React.useRef(activeTriggerId);
-  const previousPayloadRef = React.useRef(payload);
-  const pendingPayloadUpdateRef = React.useRef(false);
-
-  useIsoLayoutEffect(() => {
-    // Compare against the last committed values to decide whether we need a new DOM subtree.
-    const previousActiveTriggerId = previousActiveTriggerIdRef.current;
-    const previousPayload = previousPayloadRef.current;
-    const triggerIdChanged = activeTriggerId !== previousActiveTriggerId;
-    const payloadChanged = payload !== previousPayload;
-
-    if (triggerIdChanged) {
-      // Remount immediately on trigger change; remember if payload hasn't caught up yet.
-      setContentKey((value) => value + 1);
-      pendingPayloadUpdateRef.current = !payloadChanged;
-    } else if (pendingPayloadUpdateRef.current && payloadChanged) {
-      // Payload arrived a render later, so remount once more to avoid reusing the old <img>.
-      setContentKey((value) => value + 1);
-      pendingPayloadUpdateRef.current = false;
-    }
-
-    // Persist current values for the next render's comparison.
-    previousActiveTriggerIdRef.current = activeTriggerId;
-    previousPayloadRef.current = payload;
-  }, [activeTriggerId, payload]);
-
-  return `${activeTriggerId ?? 'current'}-${contentKey}`;
 }
